# ProyectoPlazoleta/MicroServicioPlazoleta/src/main/resources/openapi.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8082
  description: Generated server url
security:
- Bearer Authentication: []
tags:
- name: Pedidos
  description: Gestión de pedidos del sistema
- name: Platos
  description: Gestión de platos del sistema
- name: Restaurante-Empleado
  description: Gestión de relación entre restaurantes y empleados
- name: Restaurantes
  description: Gestión de restaurantes del sistema
paths:
  /restaurante/crear:
    post:
      tags:
      - Restaurantes
      summary: Crear restaurante
      description: Crea un nuevo restaurante.  ADMINISTRADOR.
      operationId: crearRestaurante
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RestauranteRequest"
        required: true
      responses:
        "201":
          description: Restaurante creado exitosamente
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RestauranteResponse"
        "400":
          description: Datos inválidos
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RestauranteResponse"
        "403":
          description: No tiene permisos para realizar esta acción
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RestauranteResponse"
  /restaurante-empleado/asignar:
    post:
      tags:
      - Restaurante-Empleado
      summary: Asignar empleado
      description: Crea la relación entre un empleado y un restaurante
      operationId: asignarEmpleado
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RestauranteEmpleadoRequest"
        required: true
      responses:
        "201":
          description: Relación creada exitosamente
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RestauranteEmpleadoResponse"
        "400":
          description: Datos inválidos
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RestauranteEmpleadoResponse"
  /plato/crear:
    post:
      tags:
      - Platos
      summary: Crear plato
      description: Crea un nuevo plato en el restaurante. PROPIETARIO.
      operationId: crearPlato
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlatoRequest"
        required: true
      responses:
        "201":
          description: Plato creado exitosamente
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PlatoResponse"
        "400":
          description: Datos inválidos
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PlatoResponse"
        "403":
          description: No tiene permisos para realizar esta acción
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PlatoResponse"
  /pedido/crear:
    post:
      tags:
      - Pedidos
      summary: Crear pedido
      description: Crea un nuevo pedido. CLIENTE.
      operationId: crearPedido
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PedidoRequest"
        required: true
      responses:
        "201":
          description: Pedido ceado exitosamente
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PedidoResponse"
        "400":
          description: Datos inválids o cliente ya tiene un pedido en proceso
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PedidoResponse"
        "403":
          description: No tiene permsos para realizar esta acción
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PedidoResponse"
  /plato/{id}:
    patch:
      tags:
      - Platos
      summary: Actualizar plato
      description: Actualiza el precio y descripción de un plato. PROPIETARIO.
      operationId: actualizarPlato
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlatoUptadeRequest"
        required: true
      responses:
        "200":
          description: Plato actualizado exitosamente
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PlatoResponse"
        "400":
          description: Datos inválidos
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PlatoResponse"
        "403":
          description: No tiene permisos para realizar esta acción
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PlatoResponse"
        "404":
          description: Plato no encontrado
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PlatoResponse"
  /plato/{id}/estado:
    patch:
      tags:
      - Platos
      summary: Cambiar estado del plato
      description: Activa o desactiva un plato. PROPIETARIO del mismo restaurante.
      operationId: cambiarEstadoPlato
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlatoEstadoRequest"
        required: true
      responses:
        "200":
          description: Estado del plato actualizado exitosamente
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PlatoResponse"
        "403":
          description: No tiene permisos para realizar esta acción
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PlatoResponse"
        "404":
          description: Plato no encontrado
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PlatoResponse"
  /pedido/{idPedido}/listo:
    patch:
      tags:
      - Pedidos
      summary: Marcar pedido como listo
      description: Marca un pedido como Listo y envía el código de entrega al cliente.
        EMPLEADO con Pedido en preparacion.
      operationId: marcarComoListo
      parameters:
      - name: idPedido
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Pedido marcado como listo exitosamente
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PedidoResponse"
        "400":
          description: El pedido no está En_Preparacion o el empleado no es el chef
            asignado
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PedidoResponse"
        "403":
          description: No tiene permisos para realizar esta acción
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PedidoResponse"
  /pedido/{idPedido}/asignar:
    patch:
      tags:
      - Pedidos
      summary: Asignar empleado a pedido
      description: Asigna el empleado autenticado a un pedido y lo pasa a En_Preparacion.
        EMPLEADO del mismo restaurante.
      operationId: asignarEmpleado_1
      parameters:
      - name: idPedido
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Empleado asignado exitosamente
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PedidoResponse"
        "400":
          description: El pedido no pertenece al restaurante del empleado
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PedidoResponse"
        "403":
          description: No tiene permisos para realizar esta acción
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PedidoResponse"
  /restaurante:
    get:
      tags:
      - Restaurantes
      summary: Listar restaurantes
      description: Lista todos los restaurantes de forma pagnada. CLIENTE.
      operationId: listarRestaurantes
      parameters:
      - name: pagina
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: tamano
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      responses:
        "200":
          description: Lista de restaurantes obtenida exitosamente
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RestauranteResponse"
        "403":
          description: No tiene permisos para realizar esta acción
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RestauranteResponse"
  /plato/restaurante/{idRestaurante}:
    get:
      tags:
      - Platos
      summary: Listar platos por restaurante
      description: Lista los platos de un restaurante con un filtro por categoría.
        CLIENTE.
      operationId: listarPlatosPorRestaurante
      parameters:
      - name: idRestaurante
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: idCategoria
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: pagina
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: tamano
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      responses:
        "200":
          description: Lista de platos obtenida exitosamente
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlatoResponse"
        "403":
          description: No tiene permisos para realizar esta acción
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlatoResponse"
  /pedido/mis-pedidos:
    get:
      tags:
      - Pedidos
      summary: Listar mis pedidos
      description: Lista todos los pedios del cliente autenticado.  CLIENTE.
      operationId: listarMisPedidos
      responses:
        "200":
          description: Lista de pedidos obtenida exitosamente
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PedidoResponse"
        "403":
          description: No tiene permisos para realizar esta acción
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PedidoResponse"
  /pedido/listar:
    get:
      tags:
      - Pedidos
      summary: Listar pedidos por estado
      description: Lista los pedidos del restaurante del empleado autenticado filtrando.
        EMPLEADO.
      operationId: listarPedidos
      parameters:
      - name: estado
        in: query
        required: true
        schema:
          type: string
      - name: pagina
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: tamano
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      responses:
        "200":
          description: Lista de pedidos obtenida exitosamente
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PedidoResponse"
        "403":
          description: No tiene permisos para realizar esta acción
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PedidoResponse"
components:
  schemas:
    RestauranteRequest:
      type: object
      properties:
        nombre:
          type: string
          minLength: 1
        nit:
          type: integer
          format: int32
        direccion:
          type: string
          minLength: 1
        telefono:
          type: string
          minLength: 1
          pattern: "^\\+?\\d{1,12}$"
        urlLogo:
          type: string
          minLength: 1
        idPropietario:
          type: integer
          format: int64
      required:
      - idPropietario
      - nit
    RestauranteResponse:
      type: object
      properties:
        nombre:
          type: string
        urlLogo:
          type: string
    RestauranteEmpleadoRequest:
      type: object
      properties:
        idEmpleado:
          type: integer
          format: int64
        idRestaurante:
          type: integer
          format: int64
      required:
      - idEmpleado
      - idRestaurante
    RestauranteEmpleadoResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        idEmpleado:
          type: integer
          format: int64
        idRestaurante:
          type: integer
          format: int64
    PlatoRequest:
      type: object
      properties:
        nombre:
          type: string
          minLength: 1
        precio:
          type: integer
          format: int64
        descripcion:
          type: string
          minLength: 1
        urlImagen:
          type: string
          minLength: 1
        idCategoria:
          type: integer
          format: int64
        idRestaurante:
          type: integer
          format: int64
      required:
      - idCategoria
      - idRestaurante
      - precio
    CategoriaResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nombre:
          type: string
        descripcion:
          type: string
    PlatoResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nombre:
          type: string
        precio:
          type: integer
          format: int64
        descripcion:
          type: string
        urlImagen:
          type: string
        idcategoria:
          $ref: "#/components/schemas/CategoriaResponse"
        activo:
          type: boolean
        idRestaurante:
          type: integer
          format: int64
    PedidoPlatoRequest:
      type: object
      properties:
        idPlato:
          type: integer
          format: int64
        cantidad:
          type: integer
          format: int32
      required:
      - cantidad
      - idPlato
    PedidoRequest:
      type: object
      properties:
        idRestaurante:
          type: integer
          format: int64
        platos:
          type: array
          items:
            $ref: "#/components/schemas/PedidoPlatoRequest"
          maxItems: 2147483647
          minItems: 1
      required:
      - idRestaurante
      - platos
    PedidoResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        idCliente:
          type: integer
          format: int64
        fecha:
          type: string
          format: date-time
        estado:
          type: string
        idRestaurante:
          type: integer
          format: int64
        platos:
          type: array
          items:
            $ref: "#/components/schemas/PedidoPlatoRequest"
    PlatoUptadeRequest:
      type: object
      properties:
        precio:
          type: integer
          format: int64
        descripcion:
          type: string
          minLength: 1
      required:
      - precio
    PlatoEstadoRequest:
      type: object
      properties:
        activo:
          type: boolean
      required:
      - activo
  securitySchemes:
    Bearer Authentication:
      type: http
      scheme: bearer
      bearerFormat: JWT